Switch to automatic registration in conatiners
Use a personal access token to allow automatic runner registration and
removal. Each runner is now installed in a LXD container and runs one
job after which the container is destroyed. The charm checks on a
configuable interval how many runners it has and creates new ones to
maintain a configured quantity.

Signed-off-by: Chris Sanders <sanders.chris@gmail.com>